SS - Safe-Systems-Tools/1.0.0/requirements.txt
ioc_finder==6.0.1
py7zr==0.11.3
rarfile==4.0
pyminizip==0.2.4
requests==2.25.1
xmltodict==0.11.0
json2xml==3.6.0
ipaddress==1.0.23
liquidpy==0.7.1
random
string
